Add Slot info management to the agents. Slot information is maintained for each deployment slot and sent to the xndeploy and control scripts.
This is also a *major* change in the way parameters are sent to the scripts, they are no longer one-by-one on the command line but through a YAML file which contains
all the parameters including the environment.

Using this galaxy branch requires a rewrite of the xndeploy and control scripts!